% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pISVM.R
\name{aaIndex}
\alias{aaIndex}
\title{aaIndex}
\usage{
aaIndex(sequence)
}
\arguments{
\item{sequence}{The peptide sequence}
}
\description{
This function compute the isoelectric point using the aa Index
}

