% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{bindRMSECorrelation}
\alias{bindRMSECorrelation}
\title{bindRMSECorrelation}
\usage{
bindRMSECorrelation(rmse, corr)
}
\arguments{
\item{correlation}{}

\item{RMSE}{}
}
\description{
This function bind the RMSE and correlation in one frame
}

