% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{bindRMSECorrelationFrame}
\alias{bindRMSECorrelationFrame}
\title{bindRMSECorrelationFrame}
\usage{
bindRMSECorrelationFrame(data, method = "pearson")
}
\arguments{
\item{data}{}

\item{method}{}
}
\description{
This function bind the RMSE and correlation in one frame
}

