% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{chargeAtPH}
\alias{chargeAtPH}
\title{chargeAtPH}
\usage{
chargeAtPH(sequence, pH = 7, pKIterative)
}
\arguments{
\item{sequence}{the aminoacid sequence}

\item{pH}{current PH}

\item{pKIterative}{the selected pK Set}
}
\description{
This fucntion compute the charge of the peptide using a pkset
}

