% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{computeCorrelationMatrix}
\alias{computeCorrelationMatrix}
\title{computeCorrelationMatrix}
\usage{
computeCorrelationMatrix(data, method = "pearson")
}
\arguments{
\item{data}{frame}

\item{method}{}
}
\description{
This function compute the correlation for the frame
}

