% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIBenchmark.R
\name{computePIvalues}
\alias{computePIvalues}
\title{computePIvalues}
\usage{
computePIvalues(originalData)
}
\arguments{
\item{originalData}{The original data frame}
}
\description{
This function take a data frame in the way of: sequence pIExp, and return a new 
data frame with all pI values computed using multiple functions. This operation could be time-consuming.
}

