% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{computeRMSEExperimental}
\alias{computeRMSEExperimental}
\title{computeRMSEExperimental}
\usage{
computeRMSEExperimental(data)
}
\arguments{
\item{dat}{}
}
\description{
This function plot the RMSE experimental for all the predicted values
}

