% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pISVM.R
\name{defaultPeptideTrainData}
\alias{defaultPeptideTrainData}
\title{defaultPeptideTrainData
This function acces to the data to trace train the SVM method}
\usage{
defaultPeptideTrainData()
}
\description{
defaultPeptideTrainData
This function acces to the data to trace train the SVM method
}

