% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pIBjell.R
\name{getcharge}
\alias{getcharge}
\title{getcharge}
\usage{
getcharge(sequence, NTermPK, CTermPK, GroupPK, pH)
}
\arguments{
\item{sequence}{}

\item{NTermPK}{the N-term pk Set}

\item{CTermPK}{the C-term pk Set}

\item{GroupPK}{the Side Group pk Set}

\item{pH}{The actual pH}
}
\description{
This function compute the charge to a SequenceAA
}

