% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPTMShift.R
\name{getCoeff}
\alias{getCoeff}
\title{getCoeff}
\usage{
getCoeff(data, xData, yData)
}
\arguments{
\item{data}{the data.frame containing data to anlysis}

\item{xData}{the x variable to anlysis}

\item{yData}{the y variable to anlysis}
}
\description{
This function retrive coefficient from a lineal model
}

