% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkSets.R
\name{loadCTermPK}
\alias{loadCTermPK}
\title{loadCTermPK}
\usage{
loadCTermPK(pkSet = "expasy")
}
\arguments{
\item{pkSet}{The pk Set to be use}
}
\description{
This function set CTerm contribution fo the Bjell method
}

