% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkSets.R
\name{loadGroupPK}
\alias{loadGroupPK}
\title{setGroupPK}
\usage{
loadGroupPK(pkSet = "expasy")
}
\arguments{
\item{pkSetMethod}{The pk Set to be use}
}
\description{
This function set the group contribution fo the Bjell method
}
\details{
#' \itemize{
  \item expasy       : http://web.expasy.org/compute_pi/pi_tool-doc.html
  \item skoog        : Skoog B. and Wichman,A. (1986) Calculation of the isoelectric points of polypeptides from the amino acid composition. Trends Anal. Chem., 5, 82–83.
  \item calibrated   : Gauci S., et al (2008) A versatile peptide pI calculator for phosphorylated and N-terminal acetylated peptides experimentally tested using peptide isoelectric focusing. Proteomics, 8, 4898–4906.
}
}

