% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkSets.R
\name{loadNTermPK}
\alias{loadNTermPK}
\title{loadNTermPK}
\usage{
loadNTermPK(pkSet = "expasy")
}
\arguments{
\item{pkSet}{The pk Set to be use}
}
\description{
This function set NTerm contribution fo the Bjell method
}

