% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pISVM.R
\name{loadSVMModel}
\alias{loadSVMModel}
\title{loadSVMModel}
\usage{
loadSVMModel(model = "default")
}
\arguments{
\item{model}{The model to be loaded}
}
\description{
This function acces to a SVM model trained
}

