% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPTMShift.R
\name{parseCoeff}
\alias{parseCoeff}
\title{parseCoeff}
\usage{
parseCoeff(coeff)
}
\arguments{
\item{value}{the value to include into any expression}
}
\description{
This function parse numerical value into formule or expression
}

