% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{pchargePhosphorylation}
\alias{pchargePhosphorylation}
\title{pchargePhosphorylation}
\usage{
pchargePhosphorylation(aa, pH)
}
\arguments{
\item{aa}{amino acid phosphorylated}

\item{pH}{current pH}
}
\description{
This pcharge function computhe the charge contribution of phosphorylation for S, T and Y amino acid
}

