% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIBjell.R
\name{pIBjell}
\alias{pIBjell}
\title{pIBjell}
\usage{
pIBjell(sequence, pkSetMethod = "expasy", gamma = 0.001)
}
\arguments{
\item{sequence}{The amino acid sequence}

\item{gamma}{The parameter to fix precision}

\item{pKSetMethod}{The pK set method}
}
\description{
This function compute the isoelectric point of ami anocid sequences using the Bjell method
}

