% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIBjell.R
\name{pIBjellMultipleSequences}
\alias{pIBjellMultipleSequences}
\title{pIBjellMultipleSequences}
\usage{
pIBjellMultipleSequences(sequences = df, pkSetMethod = "expasy")
}
\arguments{
\item{df}{The dataframe with sequences}

\item{pKSetMethod}{The pK set method}
}
\description{
This function compute the isoelectric point of sequences contained into dataframe using the Bjell method.
At lest, a column named "sequence" is expected.
It return a dataframe with predicted pI value binded.
}

