% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pICofactor.R
\name{pICofactor}
\alias{pICofactor}
\title{pICofactor}
\usage{
pICofactor(sequence)
}
\arguments{
\item{sequence}{}
}
\description{
This function compute the isoelctric following the Cofactor method proposed by
}

