% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{pIIterative}
\alias{pIIterative}
\title{pIIterative}
\usage{
pIIterative(sequence, pkSetMethod = "solomon")
}
\arguments{
\item{sequence}{The amino acid sequence to be sue to compute the Isoelectric point}

\item{pkSetMethod}{The pK set to be use from ()}
}
\description{
This function compute the isoelectric point of a sequence using the Iterative method
J Proteomics. 2012 Apr 3;75(7):2269-74. doi: 10.1016/j.jprot.2012.01.029. Epub 2012 Feb 3.
}
\examples{
pIIterative(sequence="GLPRKILCAIAKKKGKCKGPLKLVCKC", pkSetMethod = "solomon")
}

