% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{pIIterativeMultipleSequences}
\alias{pIIterativeMultipleSequences}
\title{pIIterativeMultipleSequences}
\usage{
pIIterativeMultipleSequences(sequences = df, pkSetMethod = "solomon")
}
\arguments{
\item{df}{The dataframe with sequences}

\item{pKSetMethod}{The pK set method}
}
\description{
This function compute the isoelectric point of sequences contained into dataframe using the Iterative method.
At lest, a column named "sequence" is expected.
It return a dataframe with predicted pI value binded.
}

