% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pISVM.R
\name{pISVMpeptide}
\alias{pISVMpeptide}
\title{pISVMpeptide}
\usage{
pISVMpeptide(sequence, model = "default")
}
\arguments{
\item{sequence}{The sequence to be used}

\item{model}{The SVM-based model to be used in the prediction (use "default", "heller" or "branca" options)}
}
\description{
This function predict the pI of a single sequence.
}

