% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotBoxPlotCorrelation}
\alias{plotBoxPlotCorrelation}
\title{plotBoxPlotCorrelation}
\usage{
plotBoxPlotCorrelation(dat)
}
\arguments{
\item{data}{frame}
}
\description{
This fucntion plots the raw correlation of all the predicted varaibles vs the expted variable (first column)
}

