% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotFractionCorrelation}
\alias{plotFractionCorrelation}
\title{plotFractionCorrelation}
\usage{
plotFractionCorrelation(dat)
}
\arguments{
\item{data}{frame}
}
\description{
This fucntion plots the raw correlation of all the predicted varaibles (theoretical mean of the fraction) vs the expted variable (first column). It plots error bars and confidence intervals using a lineal model.
}

