% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotHistFrecuencyValues}
\alias{plotHistFrecuencyValues}
\title{plotRawCorrelation}
\usage{
plotHistFrecuencyValues(file)
}
\arguments{
\item{file}{the file that contains the expeted value and predicted values}
}
\description{
This function plot the raw correlation between using the table variables
}

