% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotHistFunc}
\alias{plotHistFunc}
\title{plotHistFunc}
\usage{
plotHistFunc(data, na.rm = TRUE, ...)
}
\arguments{
\item{data}{dataset frame or table}

\item{na.rm}{ignore the null values}
}
\description{
This function plots the histogram for each value of the datasets, it returns a set of plots
}

