% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotOutliers.R
\name{plotOutlierDistribution}
\alias{plotOutlierDistribution}
\title{plotOutlierDistribution}
\usage{
plotOutlierDistribution(data, na.rm = TRUE, ...)
}
\arguments{
\item{data}{the data.frame that contains the expeted value and predicted values}
}
\description{
This function plot the outliers and non-outliers population distribution between using the table variables
}

