% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliers.R
\name{plotOutliers}
\alias{plotOutliers}
\title{plotOutliers}
\usage{
plotOutliers(protFile, height = 800, width = 800, cols = 3)
}
\arguments{
\item{file}{the file that contains the expeted value and predicted values}
}
\description{
This function plot the outliers and non-outliers between using the table variables
}

