% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPTMCorrelation.R
\name{plotPTMCorrelation}
\alias{plotPTMCorrelation}
\title{plotPTMCorrelation}
\usage{
plotPTMCorrelation(pepFile1, pepFile2, height = 800, width = 800,
  cols = 3)
}
\arguments{
\item{pepFile1}{with PTMs}

\item{pepFile2}{without PTMs}

\item{height}{of the chart}

\item{width}{of the chart}

\item{cols}{number of colums}
}
\description{
This function plot the PTMs correlation between to Files, one including the PTMs and another one without considering it
}

