% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotRawCorrelation}
\alias{plotRawCorrelation}
\title{plotRawCorrelation}
\usage{
plotRawCorrelation(dat)
}
\arguments{
\item{data}{frame}
}
\description{
This fucntion plots the raw correlation of all the predicted varaibles vs the expted variable (first column)
}

