% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIBenchmark.R
\docType{data}
\name{proteins}
\alias{proteins}
\title{Isoelectric poit of a set of proteins from PID-DB}
\format{A data frame with 1041 rows and 19 variables}
\usage{
data(proteins)
}
\description{
This dataset contains more than 1000 proteins with the experimental isoelectric point and the predicted pI using different methods.
The variables are as follows:
}
\details{
\itemize{
  \item PROTEIN. The potein accession
  \item EXPERIMENTAL. The experimental isoelectric point
  \item COFACTOR. The isoelectric point using the method developed by Cargile and co-workers.
  \item ITERATIVE_THURLKILL
  \item ITERATIVE_TOSELAND
  \item ITERATIVE_GRIMSLEY
  \item ITERATIVE_SOLOMON
  \item ITERATIVE_EMBOSS
  \item ITERATIVE_LEHNINGER
  \item ITERATIVE_PATRICKIOS
  \item ITERATIVE_RICHARD
  \item ITERATIVE_RODWELL
  \item ITERATIVE_SILLERO
  \item BRANCA
  \item BJELL_DEFAULT
  \item BJELL_SKOOG
  \item BJELL_CALLIBRATED
  \item BJELL_EXPASY
  \item SVM
}
}
\keyword{datasets}

