% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rawCorrelation}
\alias{rawCorrelation}
\title{plotCorrelation}
\usage{
rawCorrelation(file)
}
\arguments{
\item{file}{the file to be analyzed}
}
\description{
This function plot the raw correlation in a file: the file structure is:
}

