% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{removeFirstColumn}
\alias{removeFirstColumn}
\title{removeFirstColumn}
\usage{
removeFirstColumn(data)
}
\arguments{
\item{data}{frame or table}
}
\description{
This function remove the first in the table or frame
}

