% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkSets.R
\name{retrievePKValue}
\alias{retrievePKValue}
\title{retrievePKValue}
\usage{
retrievePKValue(aa, pKSet)
}
\arguments{
\item{aminoacid}{}

\item{pKIterative}{}
}
\description{
This function retrieve the pK values for an specific aminoacid
}

