% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(expValue, predictValue)
}
\arguments{
\item{expValue}{the value expected}

\item{predictValue}{the value predicted}
}
\description{
This function return the rmse value between to vectors
}

