% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{specify_decimal}
\alias{specify_decimal}
\title{specify_decimal}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{the double value}

\item{k}{the decimal places}
}
\description{
This function round a value to an specific decimal places
}

