% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pISVM.R
\name{svmPIBuildData}
\alias{svmPIBuildData}
\title{svmPIBuildSVMData}
\usage{
svmPIBuildData(originalData)
}
\arguments{
\item{originalData}{The original dra frame}
}
\description{
This function take a data frame in the way of: sequence pIExp. It return a dataframe with features to train a svm model
}

