% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pISVM.R
\name{svmPIBuildSVM}
\alias{svmPIBuildSVM}
\title{svmPIBuildSVMData}
\usage{
svmPIBuildSVM(originalData)
}
\arguments{
\item{originalData}{The original dra frame}
}
\description{
This function take a data frame in the way of: sequence pIExp
}

