% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pISVM.R
\name{transformData}
\alias{transformData}
\title{transformData}
\usage{
transformData(instances = instance, trainingSet = "default")
}
\arguments{
\item{instances}{The data to be transformed}

\item{trainingSet}{The training set used to apply transformation.}

\item{defaultTrainingSet}{A flag to get the training set used in the transformation (if TRUE, default training set will be used)}
}
\description{
This function transform a new data (dataframe) using center and scale transformation.
This transformation is necesary previous to use svm method.
}
\details{
If trainingSet flag is not "default", "heller" or "branca", the current data will be used to apply the transformation requeried

before use a SVM model, the set supply must contains the following variables: calibrated, expasy and aaindex.
}

