% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{bindRMSECorrelationFrame}
\alias{bindRMSECorrelationFrame}
\title{bindRMSECorrelationFrame}
\usage{
bindRMSECorrelationFrame(dat, method = "pearson")
}
\description{
This function bind the RMSE and correlation in one frame
}

