% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIBjell.R
\name{computeAllBjellValues}
\alias{computeAllBjellValues}
\title{computeAllBjellValues}
\usage{
computeAllBjellValues(seq)
}
\arguments{
\item{seq}{}
}
\description{
This function compute the isoelectric point for all the pK sets
}

