% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{computeCorrelationExperimental}
\alias{computeCorrelationExperimental}
\title{getCorrelationExperimental}
\usage{
computeCorrelationExperimental(data, method = "pearson")
}
\description{
This furntion return a matrix where each colum is a correlation respected the experimental value
}

