% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{computeCorrelationMatrix}
\alias{computeCorrelationMatrix}
\title{computeCorrelationMatrix}
\usage{
computeCorrelationMatrix(data, method = "pearson")
}
\arguments{
\item{data}{frame}
}
\description{
This function compute the correlation for the frame
}

