% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{computeRMSEExperimental}
\alias{computeRMSEExperimental}
\title{computeRMSEExperimental}
\usage{
computeRMSEExperimental(dat)
}
\description{
This function plot the RMSE experimental for all the predicted values
}

