% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkSets.R
\name{loadGroupPK}
\alias{loadGroupPK}
\title{setGroupPK}
\usage{
loadGroupPK(pkSet = "expasy")
}
\arguments{
\item{pkSetMethod}{The pk Set to be use}
}
\description{
This function set the group contribution fo the Bjell method
}

