% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{loadPkSet}
\alias{loadPkSet}
\title{loadPkSet}
\usage{
loadPkSet(pkSetMethod = "solomon")
}
\arguments{
\item{pkSetMethod}{name of the pk}
}
\description{
This function retrieve a pk dataset using the name of the pKSet
}

