% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{pIIterative}
\alias{pIIterative}
\title{pIIterative}
\usage{
pIIterative(sequence, pkSetMethod = "solomon", gamma = 0.001)
}
\arguments{
\item{sequence}{The amino acid sequence to be sue to compute the Isoelectric point}

\item{pkSetMethod}{The pK set to be used}

\item{gamma}{The parameter to fix precision}
}
\description{
This function compute the isoelectric point of a sequence using the Iterative method
Perez-Riverol et al. Isoelectric point optimization using peptide descriptors and support vector machines. J Proteomics. 2012 Apr 3;75(7):2269-74.
}
\examples{
pIIterative(sequence="GLPRKILCAIAKKKGKCKGPLKLVCKC", pkSetMethod = "solomon")

}

