% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pIIterative.R
\docType{data}
\name{pIIterativePK}
\alias{pIIterativePK}
\title{pIIterativePK}
\format{A data frame with 8 rows and 9 variables}
\usage{
data(pIIterativePK)
}
\description{
This data contains all the pk For the iterative method using different methods
}
\details{
The variables are as follows:

\itemize{
  \item AA Aminoacid
  \item SOLOMON
  \item SILLERO
  \item EMBOSS
  \item RODEWELL
  \item PATRICKIOS
  \item RICHARD
  \item LEHNINGER
}
}
\keyword{datasets}

