% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{plotCorrelationMatrix}
\usage{
plotCorrelationMatrix(data, filter = NULL)
}
\arguments{
\item{data}{DataFRame}
}
\description{
This function plot the Correlation matrix between all the columns
}

