% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pIIterative.R
\name{retrievePKValue}
\alias{retrievePKValue}
\title{retrievePKValue}
\usage{
retrievePKValue(aa, pKIterative)
}
\arguments{
\item{pKIterative}{}

\item{aminoacid}{}
}
\description{
This function retrieve the pK values for an specific aminoacid
}

